Remove all the annoying Pin stuff
I guess we can just assume that when we get the raw pointer out of a Box
it's not going to move around in weird ways. Miri isn't yelling at this
version, so I guess it can't be *that* bad, at least compared to some of
the other stuff I've done.